🚚 [variables] Move module `prompts` from `templates.domain`
