--- src/cutty/templates/adapters/questionary/__init__.py
@@ -1,1 +1,0 @@
-"""Prompts using questionary."""
