--- tests/unit/templates/adapters/questionary/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for questionary-based prompts."""
