+++ tests/unit/variables/test_prompts.py
@@ -8,9 +8,9 @@ from prompt_toolkit.input.base import PipeInput
 from prompt_toolkit.input.defaults import create_pipe_input
 from prompt_toolkit.output import DummyOutput
 
-from cutty.templates.adapters.questionary.prompts import createprompt
 from cutty.variables.binders import Binder
 from cutty.variables.bindings import GenericBinding
+from cutty.variables.prompts import createprompt
 from cutty.variables.variables import GenericVariable
 from tests.util.keys import BACKSPACE
 from tests.util.keys import ESCAPE
